pyproject.toml
[project]
name = "photodb"
version = "0.1.0"
description = "Personal photo indexing pipeline"
requires-python = ">=3.11"
dependencies = [
    "pillow>=10.0.0",
    "pillow-heif>=0.13.0",
    "click>=8.1.0",
    "python-dotenv>=1.0.0",
    "exif>=1.6.0",
    "psycopg2-binary>=2.9.10",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
process-photos = "photodb.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.black]
line-length = 100
target-version = ["py311"]
